Move the baseline's fit onto its own model screen, list its row before the first fit, and name its precision fp64 throughout.

+++ inference/src/main/kotlin/com/t1dm/inference/BaselineRunner.kt
@@ -177,7 +177,7 @@ class BaselineRunner(
             lastBg = series.mgdl[series.mgdl.size - 1],
             status = status,
             backend = BackendId.NATIVE_RIDGE_FP64,
-            precision = Precision.FP32,
+            precision = Precision.FP64,
             selected = selected,
             stale = stale,
             latencyMs = latMs,
